improved a bit localization system
on the advices of the Unity developer of the package, I improved the
loading of the localized strings.

Still have the 0.3s wait to load the strings, I will have to deal with
that. It has no problems if preloading at app startup, but more
annoying for editor scene debugging. (need adding WaitforSeconds)

Tested the SmartFormat system to translate strings depending on
conditions: formatting gender, list position, custom method...
Fun, but maybe overkill?